Enhance the Licenses page to look like Dashboard
